chore: move from "internal" to "complex" components
