+++ packages/components/src/templates/next/components/complex/SearchableTable/DGS/DGSSearchableTable.stories.tsx
