+++ packages/components/src/templates/next/components/complex/SearchableTable/Native/NativeSearchableTable.stories.tsx
