+++ packages/components/src/templates/next/components/complex/SearchableTable/Native/index.ts
