+++ packages/components/src/templates/next/components/complex/SearchableTable/SearchableTable.tsx
