+++ packages/components/src/templates/next/components/complex/SearchableTable/index.ts
