+++ packages/components/src/templates/next/components/complex/SearchableTable/shared/SearchableTableClient.tsx
@@ -4,9 +4,9 @@ import { useDeferredValue, useId, useMemo, useRef, useState } from "react"
 
 import type { SearchableTableClientProps } from "~/interfaces"
 import { tv } from "~/lib/tv"
-import BaseParagraph from "../../BaseParagraph"
-import { PaginationControls } from "../../PaginationControls"
-import { SearchField } from "../../Search"
+import BaseParagraph from "../../../internal/BaseParagraph"
+import { PaginationControls } from "../../../internal/PaginationControls"
+import { SearchField } from "../../../internal/Search"
 import { CellContent } from "./CellContent"
 import { MAX_NUMBER_OF_COLUMNS, PAGINATION_MAX_ITEMS } from "./constants"
 import { EmptyState, FallbackEmptyState } from "./EmptyState"
