+++ packages/components/src/templates/next/components/complex/SearchableTable/shared/constants.ts
