+++ packages/components/src/templates/next/components/complex/index.ts
@@ -18,3 +18,4 @@ export { CollectionBlock } from "./CollectionBlock"
 export { ImageGallery } from "./ImageGallery"
 export { default as ChildrenPages } from "./ChildrenPages"
 export { LogoCloud } from "./LogoCloud"
+export { SearchableTable } from "./SearchableTable"
