+++ packages/components/src/templates/next/components/internal/index.ts
@@ -13,7 +13,6 @@ export { default as Navbar } from "./Navbar"
 export { default as Notification } from "./Notification"
 export { PaginationControls } from "./PaginationControls"
 export { default as Pill } from "./Pill"
-export { SearchableTable } from "./SearchableTable"
 export { default as SearchSGInputBox } from "./SearchSGInputBox"
 export { default as Siderail } from "./Siderail"
 export { SkipToContent } from "./SkipToContent"
